backend/app/requirements.txt
# app/requirements.txt
ultralytics
Pillow
uvicorn
fastapi
python-multipart
gunicorn